refactor: Update folder structure for explore pages
