+++ frontend/js/src/explore/art-creator/components/ColorPicker.tsx
