+++ frontend/js/src/explore/art-creator/components/IconTray.tsx
