+++ frontend/js/src/explore/art-creator/components/Preview.tsx
