+++ frontend/js/src/explore/fresh-releases/components/ReleaseTimeline.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import Slider from "rc-slider";
 import { countBy, debounce, zipObject } from "lodash";
-import { formatReleaseDate, useMediaQuery } from "./utils";
+import { formatReleaseDate, useMediaQuery } from "../utils";
 
 type ReleaseTimelineProps = {
   releases: Array<FreshReleaseItem>;
