+++ frontend/js/src/explore/lb-radio/components/Playlist.tsx
@@ -4,8 +4,8 @@ import { IconProp } from "@fortawesome/fontawesome-svg-core";
 import { faCog } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import React from "react";
-import PlaylistItemCard from "../../playlists/PlaylistItemCard";
-import PlaylistMenu from "../../playlists/PlaylistMenu";
+import PlaylistItemCard from "../../../playlists/PlaylistItemCard";
+import PlaylistMenu from "../../../playlists/PlaylistMenu";
 
 type LBRadioFeedbackProps = {
   feedback: string[];
