+++ frontend/js/src/explore/year-in-music/2021/components/ComponentToImage.tsx
@@ -3,9 +3,9 @@ import React, { useState } from "react";
 import { faCamera, faHeadphones } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { IconProp } from "@fortawesome/fontawesome-svg-core";
-import ListenCard from "../../../listens/ListenCard";
-import { getEntityLink } from "../../../stats/utils";
-import Loader from "../../../components/Loader";
+import ListenCard from "../../../../listens/ListenCard";
+import { getEntityLink } from "../../../../stats/utils";
+import Loader from "../../../../components/Loader";
 
 export type ComponentToImageProps = {
   data: any[];
