+++ frontend/js/src/explore/year-in-music/2021/page.tsx
@@ -27,7 +27,7 @@ import withAlertNotifications from "../../../notifications/AlertNotificationsHOC
 
 import { getPageProps } from "../../../utils/utils";
 import { getEntityLink } from "../../../stats/utils";
-import ComponentToImage from "./ComponentToImage";
+import ComponentToImage from "./components/ComponentToImage";
 
 import ListenCard from "../../../listens/ListenCard";
 import UserListModalEntry from "../../../follow/UserListModalEntry";
