+++ frontend/js/src/explore/year-in-music/2022/components/MagicShareButton.tsx
