+++ webpack.config.js
@@ -67,28 +67,28 @@ module.exports = function (env, argv) {
       userFeed: path.resolve(jsDir, "src/user-feed/UserFeed.tsx"),
       playlist: path.resolve(jsDir, "src/playlists/Playlist.tsx"),
       playlists: path.resolve(jsDir, "src/playlists/Playlists.tsx"),
-      explore: path.resolve(jsDir, "src/explore/ExploreCard.tsx"),
-      huesound: path.resolve(jsDir, "src/explore/huesound/ColorPlay.tsx"),
-      lb_radio: path.resolve(jsDir, "src/explore/lb-radio/LBRadio.tsx"),
+      explore: path.resolve(jsDir, "src/explore/page.tsx"),
+      huesound: path.resolve(jsDir, "src/explore/huesound/page.tsx"),
+      lb_radio: path.resolve(jsDir, "src/explore/lb-radio/page.tsx"),
       yearInMusic2021: path.resolve(
         jsDir,
-        "src/explore/year-in-music/2021/YearInMusic.tsx"
+        "src/explore/year-in-music/2021/page.tsx"
       ),
       yearInMusic2022: path.resolve(
         jsDir,
-        "src/explore/year-in-music/2022/YearInMusic.tsx"
+        "src/explore/year-in-music/2022/page.tsx"
       ),
       yearInMusic2023: path.resolve(
         jsDir,
-        "src/explore/year-in-music/2023/YearInMusic.tsx"
+        "src/explore/year-in-music/2023/page.tsx"
       ),
       coverArtComposite2022: path.resolve(
         jsDir,
-        "src/explore/year-in-music/2022/CoverArtComposite.tsx"
+        "src/explore/cover-art-collage/2022/page.tsx"
       ),
       coverArtComposite2023: path.resolve(
         jsDir,
-        "src/explore/year-in-music/2023/CoverArtComposite.tsx"
+        "src/explore/cover-art-collage/2023/page.tsx"
       ),
       homepage: path.resolve(jsDir, "src/home/Homepage.tsx"),
       recommendationsPlayground: path.resolve(
@@ -108,10 +108,7 @@ module.exports = function (env, argv) {
         jsDir,
         "src/metadata-viewer/MetadataViewerPageWrapper.tsx"
       ),
-      freshReleases: path.resolve(
-        jsDir,
-        "src/explore/fresh-releases/FreshReleases.tsx"
-      ),
+      freshReleases: path.resolve(jsDir, "src/explore/fresh-releases/page.tsx"),
       selectTimezone: path.resolve(
         jsDir,
         "src/user-settings/SelectTimezone.tsx"
@@ -120,10 +117,7 @@ module.exports = function (env, argv) {
         jsDir,
         "src/user-settings/SelectTroiPreferences.tsx"
       ),
-      artCreator: path.resolve(
-        jsDir,
-        "src/explore/stats-art-designer/ArtCreator.tsx"
-      ),
+      artCreator: path.resolve(jsDir, "src/explore/art-creator/page.tsx"),
       artistPage: path.resolve(jsDir, "src/entity-pages/ArtistPage.tsx"),
       albumPage: path.resolve(jsDir, "src/entity-pages/AlbumPage.tsx"),
     },
